Improved code quality

+++ master-ugr.calculator.lib/calculator.lib/NumberAttributter.cs
@@ -32,7 +32,7 @@ namespace calculator.lib
 
         public static double Sqrt(int number)
         {
-            return number < 0 ? double.NaN : (double)Math.Round(Math.Sqrt(number),2);
+            return number < 0 ? double.NaN : Math.Round(Math.Sqrt(number),2);
         }
     }
 }
